# Improve linear indexing performance for FastSubArrays