--- a/base/subarray.jl
+++ b/base/subarray.jl
@@ -314,12 +314,8 @@ end
 function getindex(V::FastContiguousSubArray, i::AbstractUnitRange{Int})
     @inline
     @boundscheck checkbounds(V, i)
-    out = similar(V, axes(i))
-    li = length(i)
-    if li > 0
-        copyto!(out, firstindex(out), V.parent, V.offset1 + first(i), li)
-    end
-    return out
+    @inbounds r = V.parent[V.offset1 .+ i]
+    r
 end
 
 # For vector views with linear indexing, we disambiguate to favor the stride/offset
@@ -335,16 +331,6 @@ function getindex(V::FastContiguousSubArray{<:Any, 1}, i::Int)
     @boundscheck checkbounds(V, i)
     @inbounds r = V.parent[V.offset1 + i]
     r
-end
-function getindex(V::FastContiguousSubArray{<:Any, 1}, i::AbstractUnitRange{Int})
-    @inline
-    @boundscheck checkbounds(V, i)
-    out = similar(V, axes(i))
-    li = length(i)
-    if li > 0
-        copyto!(out, firstindex(out), V.parent, V.offset1 + first(i), li)
-    end
-    return out
 end
 @inline getindex(V::FastContiguousSubArray, i::Colon) = getindex(V, to_indices(V, (:,))...)
 
@@ -367,13 +353,16 @@ function setindex!(V::FastContiguousSubArray, x, i::Int)
     @inbounds V.parent[V.offset1 + i] = x
     V
 end
+function setindex!(V::FastSubArray, x, i::AbstractUnitRange{Int})
+    @inline
+    @boundscheck checkbounds(V, i)
+    @inbounds V.parent[V.offset1 .+ V.stride1 .* i] = x
+    V
+end
 function setindex!(V::FastContiguousSubArray, x, i::AbstractUnitRange{Int})
     @inline
     @boundscheck checkbounds(V, i)
-    li = length(i)
-    if li > 0
-        copyto!(V.parent, V.offset1 + first(i), x, firstindex(x), li)
-    end
+    @inbounds V.parent[V.offset1 .+ i] = x
     V
 end
 
@@ -383,25 +372,10 @@ function setindex!(V::FastSubArray{<:Any, 1}, x, i::Int)
     @inbounds V.parent[V.offset1 + V.stride1*i] = x
     V
 end
-function setindex!(V::FastSubArray{<:Any, 1}, x, i::AbstractUnitRange{Int})
-    @inline
-    @boundscheck checkbounds(V, i)
-    @inbounds V.parent[V.offset1 .+ V.stride1 .* i] = x
-    V
-end
 function setindex!(V::FastContiguousSubArray{<:Any, 1}, x, i::Int)
     @inline
     @boundscheck checkbounds(V, i)
     @inbounds V.parent[V.offset1 + i] = x
-    V
-end
-function setindex!(V::FastContiguousSubArray{<:Any, 1}, x, i::AbstractUnitRange{Int})
-    @inline
-    @boundscheck checkbounds(V, i)
-    li = length(i)
-    if li > 0
-        copyto!(V.parent, V.offset1 + first(i), x, firstindex(x), li)
-    end
     V
 end
 @inline setindex!(V::FastSubArray, x, i::Colon) = setindex!(V, x, to_indices(V, (i,))...)