@johnnychen94 2022-05-20
> There are a lot of lines changes in https://github.com/JuliaLang/julia/commit/4232be008751de9808593599aeb2e48b39cb102f, but that's mainly to get the OffsetArrays test helper in sync with the repo.

I guess it's better to open a separate PR for this and merge it quickly (since it's only a test dependency). I left the version note here so we need to update it as well https://github.com/JuliaLang/julia/blob/8230203c8cf187ed15d0963314502f2f604d17a3/test/testhelpers/OffsetArrays.jl#L8