@N5N3 2022-05-20
I'm suppressing that LLVM fails to vectorlize the copy part of genetic `getindex`/`setindex!` even for 1d case.
Maybe we can optimize them if `getindex`'s `dest` (or `setindex!`'s `src`) has linear index.